mapping function
these mapping functions would take in scraped data such as the example implementation in scraper.py (the data with which was imported from beautifulsoup) and will map each to the standardized data we use. could edit the mapping function as different files have different fields that map to the same thing, and build a database of it.