mock functions
can map the list of tuples to this data and use json.dumps at the very end on a list of the json data points to save the final result in the "data" folder. due to time scarcity I was unable to implement this but it works exactly like in the 'scraper.py' file. mine outputs as a json file but a better implementation would have it as a list / txt file for easy function access and then output the final draft as a json.