Not complete
Was unable to reference every function on data due to lack of time but ideally I will have one function in both cleaning and standardize that call all of the functions and input the list from the scraped data. It will then be saved into 'data' as a json file. This is my first time implementing something like this on this scale with such a time frame, however, the scraper works which is the difficult part. With another several days I would be able to implement the proper functions and end to end as described.